Making Vulkan API version default to 1.1.0
There is some bug in spv generated by clspv. The generated spv is corrected by removing invalid code by reverse engineering. Version upgrade is required to make this hack work.
>>>First generate the spv by compiling, then disassemble it.
As per the error, there is one variable which has NonWritable property associated with it which it doesn't expect. So, removed line with `NonWritable`.
Reassambled the spv.
Changed the Vulkan version to 1.1